requirements.txt
snowflake-connector-python
snowflake-snowpark-python
python-dotenv
pyyaml
pytest
pytest-mock
pydantic-settings
